added composer require packages and also changed the migration from integer to biginteger

+++ src/Database/Migrations/2018_03_22_054642_create_dz_column_definitions_table.php
@@ -16,14 +16,14 @@ class CreateDzColumnDefinitionsTable extends Migration {
         Schema::create('dz_column_definitions', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('name');
             $table->string('description')->nullable();
             $table->string('supported_identifiers')->nullable();
 
-            $table->unsignedInteger('created_by')->nullable();
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
 
             $table->foreign('created_by')->references('id')->on($userTable);
             $table->foreign('updated_by')->references('id')->on($userTable);
