+++ src/Database/Migrations/2018_05_22_130109_create_dz_form_preferences_table.php
@@ -15,18 +15,18 @@ class CreateDzFormPreferencesTable extends Migration {
         Schema::create('dz_form_preferences', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('source_type')->nullable();
-            $table->unsignedInteger('source_id')->nullable();
+            $table->unsignedBigInteger('source_id')->nullable();
 
             $table->string('name')->nullable();
             $table->string('identifier')->nullable();
 
             $table->text('column_definition')->nullable();
 
-            $table->unsignedInteger('created_by')->nullable();
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
 
             $table->foreign('created_by')->references('id')->on($userTable);
             $table->foreign('updated_by')->references('id')->on($userTable);
