+++ src/Database/Migrations/2019_03_07_052833_create_dz_observer_rules_table.php
@@ -15,21 +15,21 @@ class CreateDzObserverRulesTable extends Migration {
         Schema::create('dz_observer_rules', function (Blueprint $table) {
             $userTable = LaravelUtility::getUserTable();
 
-            $table->increments('id');
+            $table->bigIncrements('id');
 
             $table->string('name');
             $table->string('description');
 
-            $table->unsignedInteger('model_id')->nullable();
-            $table->unsignedInteger('trigger_type_id')->nullable();
+            $table->unsignedBigInteger('model_id')->nullable();
+            $table->unsignedBigInteger('trigger_type_id')->nullable();
 
-            $table->unsignedInteger('execution_order')->default(1);
+            $table->unsignedBigInteger('execution_order')->default(1);
             $table->boolean('active')->default(true);
 
             $table->text('filter_condition')->nullable();
 
-            $table->unsignedInteger('created_by')->nullable();
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
+            $table->unsignedBigInteger('updated_by')->nullable();
 
             $table->foreign('model_id')->references('id')->on('dz_model_details');
             $table->foreign('trigger_type_id')->references('id')->on('dz_lookup_values');
