+++ src/Database/Migrations/2019_05_03_090045_remove_dz_observer_actions_observer_rule_id_table.php
@@ -24,7 +24,7 @@ class RemoveDzObserverActionsObserverRuleIdTable extends Migration {
      */
     public function down () {
         Schema::table('dz_observer_actions', function (Blueprint $table) {
-            $table->unsignedInteger('observer_rule_id')->nullable();
+            $table->unsignedBigInteger('observer_rule_id')->nullable();
             $table->foreign('observer_rule_id')->references('id')->on('dz_lookup_values');
         });
     }
